InfiniteScrollItems transformed into withInfiniteScroll
